still refactoring

+++ Lib/DataTypes/MonteCarlo/LifetimeSpend.cs
@@ -2,10 +2,10 @@ namespace Lib.DataTypes.MonteCarlo;
 
 public class LifetimeSpend
 {
-    public long TotalSpendLifetime { get; set; } = 0;
-    public long TotalInvestmentAccrualLifetime { get; set; } = 0;
-    public long TotalDebtAccrualLifetime { get; set; } = 0;
-    public long TotalSocialSecurityWageLifetime { get; set; } = 0;
-    public long TotalDebtPaidLifetime { get; set; } = 0;
+    public decimal TotalSpendLifetime { get; set; } = 0m;
+    public decimal TotalInvestmentAccrualLifetime { get; set; } = 0m;
+    public decimal TotalDebtAccrualLifetime { get; set; } = 0m;
+    public decimal TotalSocialSecurityWageLifetime { get; set; } = 0m;
+    public decimal TotalDebtPaidLifetime { get; set; } = 0m;
     public bool IsBankrupt { get; set; } = false;
 }
