+++ Lib/DataTypes/MonteCarlo/McDebtPosition.cs
@@ -7,8 +7,7 @@ public record McDebtPosition
     public required bool IsOpen { get; set; } = true;
     public required string Name { get; set; }
     public required LocalDateTime Entry { get; set; }
-    public required long AnnualPercentageRate { get; set; }
-    public required long MonthlyPayment { get; set; }
-    public required long CurrentBalance { get; set; }
+    public required decimal AnnualPercentageRate { get; set; }
+    public required decimal MonthlyPayment { get; set; }
+    public required decimal CurrentBalance { get; set; }
 }
-
